--- app/src/main/res/layout/fragment_dashboard.xml
@@ -1,60 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.core.widget.NestedScrollView
-        xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:app="http://schemas.android.com/apk/res-auto"
-        xmlns:tools="http://schemas.android.com/tools"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="#20aaaaaa"
-        tools:context=".ui.dashboard.DashboardFragment">
-
-
-<!--        <com.google.android.material.imageview.ShapeableImageView-->
-<!--            android:id="@+id/my_image"-->
-<!--            android:layout_width="108dp"-->
-<!--            android:layout_height="108dp"-->
-<!--            android:layout_marginTop="102dp"-->
-<!--            tools:src="@tools:sample/avatars"-->
-<!--            app:strokeColor="@android:color/darker_gray"-->
-<!--            app:strokeWidth="4dp"-->
-<!--            app:shapeAppearance="@style/ShapeAppearanceOverlay.App.CornerSize50Percent"-->
-<!--            app:layout_constraintTop_toTopOf="parent"-->
-<!--            app:layout_constraintStart_toStartOf="parent"-->
-<!--            app:layout_constraintEnd_toEndOf="parent"/>-->
-
-<!--        <androidx.appcompat.widget.AppCompatTextView-->
-<!--            android:id="@+id/my_nickname"-->
-<!--            android:layout_width="wrap_content"-->
-<!--            android:layout_height="wrap_content"-->
-<!--            android:layout_marginTop="24dp"-->
-<!--            android:textStyle="bold"-->
-<!--            android:textSize="24sp"-->
-<!--            android:textColor="@android:color/black"-->
-<!--            app:layout_constraintTop_toBottomOf="@id/my_image"-->
-<!--            app:layout_constraintStart_toStartOf="parent"-->
-<!--            app:layout_constraintEnd_toEndOf="parent"-->
-<!--            tools:text="@tools:sample/full_names"/>-->
-
-<!--        <LinearLayout-->
-<!--            android:layout_width="match_parent"-->
-<!--            android:layout_height="wrap_content"-->
-<!--            android:background="@android:color/white"-->
-<!--            android:layout_marginTop="18dp"-->
-<!--            android:orientation="vertical"-->
-<!--            app:layout_constraintTop_toBottomOf="@id/my_nickname">-->
-
-<!--            <include-->
-<!--                android:id="@+id/my_favorite_list"-->
-<!--                layout="@layout/view_my_item"/>-->
-<!--            <include-->
-<!--                layout="@layout/view_my_item"/>-->
-<!--            <include-->
-<!--                layout="@layout/view_my_item"/>-->
-<!--        </LinearLayout>-->
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</androidx.core.widget.NestedScrollView>
