+++ app/src/main/res/layout/fragment_mypage.xml
@@ -10,7 +10,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="#20aaaaaa"
-        tools:context=".ui.dashboard.DashboardFragment">
+        tools:context=".ui.home.HomeFragment">
 
 
         <com.google.android.material.imageview.ShapeableImageView
