+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -7,7 +7,7 @@
         android:title="@string/title_home" />
 
     <item
-        android:id="@+id/feed"
+        android:id="@+id/navigation_feed"
         android:icon="@drawable/ic_feed"
         android:title="@string/title_dashboard" />
