+++ app/src/main/res/navigation/main_navigation.xml
@@ -8,22 +8,19 @@
     <fragment
         android:id="@+id/navigation_home"
         android:name="com.gdg.withtrip.ui.home.HomeFragment"
-        android:label="@string/title_home"
-        tools:layout="@layout/fragment_home">
+        android:label="@string/title_home">
+
+    </fragment>
+
+    <fragment
+        android:id="@+id/navigation_feed"
+        android:name="com.gdg.withtrip.ui.feed.FeedFragment"
+        android:label="@string/title_dashboard"
+        tools:layout="@layout/fragment_feed">
 
         <action
             android:id="@+id/navigate_to_trip_detail"
             app:destination="@id/tripDetailFragment" />
-    </fragment>
-
-    <fragment
-        android:id="@+id/navigation_dashboard"
-        android:name="com.gdg.withtrip.ui.dashboard.DashboardFragment"
-        android:label="@string/title_dashboard"
-        tools:layout="@layout/fragment_dashboard" >
-        <action
-            android:id="@+id/action_mainFragment_to_likeFragment"
-            app:destination="@id/likeFragment" />
     </fragment>
 
     <fragment
@@ -34,5 +31,10 @@
     <fragment
         android:id="@+id/navigation_mypage"
         android:name="com.gdg.withtrip.ui.mypage.MyPageFragment"
-        android:label="MyPageFragment" />
+        android:label="MyPageFragment" >
+        <action
+            android:id="@+id/action_mypageFragment_to_likeFragment"
+            app:destination="@id/likeFragment" />
+
+    </fragment>
 </navigation>
